[Divide by 0 Error] lu check migrate to c++
