+++ python/paddle/fluid/tests/unittests/test_lu_op.py
@@ -303,6 +303,20 @@ class TestLUAPI(unittest.TestCase):
             run_lu_static(tensor_shape, dtype)
 
 
+class TestLUAPIError(unittest.TestCase):
+    def test_errors(self):
+        with paddle.fluid.dygraph.guard():
+            # The size of input in lu should not be 0.
+            def test_0_size():
+                array = np.array([], dtype=np.float32)
+                x = paddle.to_tensor(
+                    np.reshape(array, [0, 0, 0]), dtype='float32'
+                )
+                paddle.linalg.lu(x, get_infos=True)
+
+            self.assertRaises(ValueError, test_0_size)
+
+
 if __name__ == "__main__":
     paddle.enable_static()
     unittest.main()
